Correct Brat Green Hex colour code
On the site https://www.bratgreen.com/ the code is shown as #8ACE00 on the page, but using the "copy" button actually puts #89CC04 onto the clipboard, which is what went onto the Jira ticket.  These are visually very close, but it's nice to use the canonical hex value for the colour, and this will also give parity with the main register frontend services.